% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{cartesian_to_euler_covariance}
\alias{cartesian_to_euler_covariance}
\title{Convert Cartesian Covariance Matrix to Euler Angle Covariance Matrix}
\usage{
cartesian_to_euler_covariance(q)
}
\arguments{
\item{q}{Unit quaterion.}
}
\value{
H matrix (3x3) to properly transform cartesian covariance matrix to euler covariance matrix.
}
\description{
Convert Cartesian Covariance Matrix to Euler Angle Covariance Matrix
}
