% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{euler_to_quaterion}
\alias{euler_to_quaterion}
\title{Euler angles to quaternion of rotation}
\usage{
euler_to_quaterion(pitch, roll, yaw)
}
\arguments{
\item{pitch, roll, yaw}{euler angles}
}
\value{
quaternion of rotation
}
\description{
Euler angles to quaternion of rotation
}
