% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{euler_to_rotation}
\alias{euler_to_rotation}
\title{Get rotation matrix based on provided pitch, roll and yaw angles.}
\usage{
euler_to_rotation(pitch, roll, yaw, order = "xyz", units = "degrees")
}
\arguments{
\item{pitch, roll, yaw}{euler angles of rotation.}

\item{order}{order of rotation. only currently takes "xyz"}

\item{units}{units for angles: degrees or radians.}
}
\value{
Rotation matrix.
}
\description{
Get rotation matrix based on provided pitch, roll and yaw angles.
}
