% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{find_rotation}
\alias{find_rotation}
\title{Find optimal rotation}
\usage{
find_rotation(X, Y, method = "q", output = "euler", sd = NA)
}
\arguments{
\item{X}{matrix of vectors to be aligned with Y}

\item{Y}{matrix of vectors}

\item{method}{method for finding optimal rotation The default is Davenport's q-method.}

\item{output}{format for rotation. Output can be set to: (1) "euler" (pitch, roll, yaw), (2) "rotation" (3x3 rotation matrix) or (3) quaternion (rotation quaternion). Defaults to euler angles.}

\item{sd}{standard deviation}
}
\value{


Rotation in the form request by output. Returns covariance estimate is \code{sd} is provided.
}
\description{
Find optimal rotation
}
