% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{get_covariance}
\alias{get_covariance}
\title{Find Covariance Matrix for Rotation}
\usage{
get_covariance(X, Y, units = "euler", sd = NA)
}
\arguments{
\item{X}{Matrix of vector observations}

\item{Y}{Matrix of vector observations}
}
\value{
Covariance matrix
}
\description{
Find Covariance Matrix for Rotation
}
