% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{quaterion_to_euler}
\alias{quaterion_to_euler}
\title{Quaterion to euler angles}
\usage{
quaterion_to_euler(q)
}
\arguments{
\item{q}{quaternion of rotation.}
}
\value{
sequence of euler angles specifying rotation.
}
\description{
Quaterion to euler angles
}
