% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{rotation_to_euler}
\alias{rotation_to_euler}
\title{Converts rotation matrix to euler angles.}
\usage{
rotation_to_euler(R)
}
\arguments{
\item{R}{rotation matrix}
}
\value{
sequence of euler angles specifying rotation.
}
\description{
Converts rotation matrix to euler angles.
}
