% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{rotation_to_quaterion}
\alias{rotation_to_quaterion}
\title{Rotation matrix to quaternion of rotation.}
\usage{
rotation_to_quaterion(R)
}
\arguments{
\item{R}{rotation matrix}
}
\value{
quaternion of rotation.
}
\description{
Rotation matrix to quaternion of rotation.
}
